+++ .storybook/theme.js
@@ -1,8 +1,11 @@
-import { create } from '@storybook/theming';
+import {create} from '@storybook/theming';
 
 export default create({
-  brandTitle: 'The good store',
-  brandUrl: 'https://home-thegoodstore.netlify.app/',
-  brandImage: 'https://res.cloudinary.com/dlwdq84ig/image/upload/w_1920,q_auto,c_scale/dzseuok3zjub8o70bgko',
+  base: 'light',
+  brandTitle: 'B2C Retail Store',
+  brandUrl: 'https://poc-b2cdev.netlify.app/',
+  brandImage: 'https://res.cloudinary.com/dlwdq84ig/image/upload/v1715784695/ukxoguhuiveasua5bbkg.png',
   brandTarget: '_self',
+  appBg: '#111152',
 });
+