+++ components/commercetools-ui/organisms/product/product-list/components/list/index.tsx
@@ -1,9 +1,9 @@
 import React, { useMemo } from 'react';
 import { Product } from 'shared/types/product/Product';
-import ProductTile from 'components/commercetools-ui/organisms/product/product-tile';
 import { PLP_PRODUCT_CLICKED } from 'helpers/constants/events';
 import { useFormat } from 'helpers/hooks/useFormat';
 import { useProductList } from '../../context';
+import PlpProductTileWrapper from '../plp-product-tile-wrapper';
 
 interface Props {
   products: Product[];
@@ -20,7 +20,7 @@ const List: React.FC<Props> = ({ products }) => {
     <div>
       <div className="grid grid-cols-2 gap-16 pt-32 md:grid-cols-3 lg:grid-cols-4 lg:gap-24">
         {products.map((product) => (
-          <ProductTile
+          <PlpProductTileWrapper
             key={product.productId}
             product={product}
             isSearchResult={!!searchQuery}