+++ frontastic/renderer/utils/highlight/index.ts
@@ -1,3 +1,3 @@
 export const highlight = (isHighlighted?: boolean) => {
-  return isHighlighted ? 'outline-accent-red outline outline-dashed outline-2' : '';
+  return isHighlighted ? 'border-accent-red border border-dashed border-2' : '';
 };